src/recorder/requirements.txt
boto3>=1.28.52
python-dotenv>=1.0.0
Pillow>=10.0.1
pyautogui>=0.9.54